bug: 22362
Handle MailItems representing items larger than 2GB.  Increase the MAIL_ITEM.SIZE and REVISION.SIZE database columns to BIGINT.  Change MailItem.mData.size and MailItem.getSize() to return a long.  Tweak a few things around the app to accommodate this, with the only sketchy bit being that the methods returning item content in a byte[] now must throw MailServiceException.MESSAGE_TOO_BIG on >2GB items because Java doesn't support byte arrays larger than that.

Add Folder.getItemCount() and try to call it instead of calling Folder.getSize() when we want to know the number of non-subfolder items in a folder.

DB version number is now 50.  Please reset-the-world or run the migrate20071206-WidenSizeColumns.pl migration script to get your install's database schema in sync.

http://bugzilla.zimbra.com/show_bug.cgi?id=22362

+++ ZimbraOffline/src/db/create_database.sql
@@ -28,7 +28,7 @@ CREATE TABLE IF NOT EXISTS ${DATABASE_NAME}.mail_item (
    index_id      INTEGER UNSIGNED,
    imap_id       INTEGER UNSIGNED,
    date          INTEGER UNSIGNED NOT NULL,  -- stored as a UNIX-style timestamp
-   size          INTEGER UNSIGNED NOT NULL,
+   size          BIGINT UNSIGNED NOT NULL,
    volume_id     TINYINT UNSIGNED,
    blob_digest   VARCHAR(28) BINARY,         -- SHA-1 hash of blob, or NULL if item has no blob
    unread        INTEGER UNSIGNED,           -- stored separately from the other flags so we can index it
@@ -69,7 +69,7 @@ CREATE TABLE IF NOT EXISTS ${DATABASE_NAME}.revision (
    item_id       INTEGER UNSIGNED NOT NULL,
    version       INTEGER UNSIGNED NOT NULL,
    date          INTEGER UNSIGNED NOT NULL,  -- stored as a UNIX-style timestamp
-   size          INTEGER UNSIGNED NOT NULL,
+   size          BIGINT UNSIGNED NOT NULL,
    volume_id     TINYINT UNSIGNED,
    blob_digest   VARCHAR(28) BINARY,         -- SHA-1 hash of blob, or NULL if item has no blob
    name          VARCHAR(128),               -- namespace entry for item (e.g. tag name, folder name, document filename)
